pyproject.toml
[project]
name = "arbos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "python-dotenv",
    "pyTelegramBotAPI",
    "torch>=2.10.0",
    "numpy>=2.2.6",
    "basilica-sdk>=0.25.0",
    "fastapi",
    "uvicorn",
]

[tool.setuptools]
py-modules = ["arbos"]
